Move 3 monitor-related methods from Win32Window to BaseWindow
These functions should work with each Window implementation once
the Monitor classes are implemented.

This commit also adds docstrings for each method.
