+++ dragonfly/windows/base_window.py
@@ -24,9 +24,12 @@ Base Window class
 
 """
 
-from six import string_types, integer_types, binary_type
 from locale import getpreferredencoding
 
+from six import string_types, integer_types, binary_type
+
+from .monitor import monitors
+from .rectangle import unit
 from .window_movers import window_movers
 
 #===========================================================================
@@ -239,6 +242,53 @@ class BaseWindow(object):
         :type rectangle: Rectangle
         """
         raise NotImplementedError()
+
+    def get_containing_monitor(self):
+        """
+        Method to get the :class:`Monitor` containing the window.
+
+        This checks which monitor contains the center of the window.
+
+        :returns: containing monitor
+        :rtype: Monitor
+        """
+        center = self.get_position().center
+        for monitor in monitors:
+            if monitor.rectangle.contains(center):
+                return monitor
+        # Fall through, return first monitor.
+        return monitors[0]
+
+    def get_normalized_position(self):
+        """
+        Method to get the window's normalized position.
+
+        This is useful when working with multiple monitors.
+
+        :returns: normalized position
+        :rtype: Rectangle
+        """
+        monitor = self.get_containing_monitor()
+        rectangle = self.get_position()
+        rectangle.renormalize(monitor.rectangle, unit)
+        return rectangle
+
+    def set_normalized_position(self, rectangle, monitor=None):
+        """
+        Method to get the window's normalized position.
+
+        This is useful when working with multiple monitors.
+
+        :param rectangle: window position
+        :type rectangle: Rectangle
+        :param monitor: monitor to normalize to (default: the first one).
+        :type monitor: Monitor
+        """
+        if not monitor:
+            monitor = self.get_containing_monitor()
+
+        rectangle.renormalize(unit, monitor.rectangle)
+        self.set_position(rectangle)
 
     #-----------------------------------------------------------------------
     # Methods for miscellaneous window control.
